API rev update, fix IOPS compare and Snapsched.
This commit ups the API revision from 7.0 to 8.0 in anticipation of the
snapshed module, which isn't avaliable in 7.0. Adds munge to the iop
resources so they compare properly. Lastly add a get schedule by name method
to the api to make up for lack of it in the native API.