Change the solidfire_vag properties to set an array of volume names
Currently, the properties on a solidfire_vag are a list of volume IDs
whereas the property that gets set in the config is a list of volume
names.  This causes a diff every time the module runs.  This changes the
property to get a list of names instead of IDs so there is only a diff
if the volume array changes.